requirements.txt
opencv-contrib-python 
numpy 
Pillow
cvzone
mediapipe